feat: move caching-service to api-server
