+++ core/api-server/lib/service/caching.js
@@ -1,29 +1,77 @@
-const request = require('request-promise');
-const querystring = require('querystring');
+const { NodesMap: DAG } = require('@hkube/dag');
+const { parser, consts } = require('@hkube/parsers');
+const stateManager = require('../state/state-manager');
+const { InvalidDataError, } = require('../errors');
+const { relations } = consts;
 
 class CachingService {
-    init(config) {
-        const { protocol, host, port, prefix } = config.cachingServer;
-        this._baseUrl = `${protocol}://${host}:${port}/${prefix}`;
+    async exec({ jobId, nodeName }) {
+        const stored = await this._getStoredExecution(jobId);
+        this._validateType(stored.nodes);
+        const { successors } = this._findRelations(stored, nodeName);
+        const pipeline = this._createSubPipeline(stored, nodeName, successors);
+        return pipeline;
     }
 
-    async exec(options) {
-        const { jobId, nodeName } = options;
-        const qs = querystring.stringify({ jobId, nodeName });
-        const uri = `${this._baseUrl}?${qs}`;
-        let error;
-        let pipeline;
-        try {
-            pipeline = await request({
-                method: 'GET',
-                uri,
-                json: true
+    _validateType(nodes) {
+        const node = nodes.find(n => parser.findNodeRelation(n.input, relations.WAIT_ANY));
+        if (node) {
+            throw new InvalidDataError(`relation ${relations.WAIT_ANY} for node ${node.nodeName} is not allowed`);
+        }
+    }
+
+    _createSubPipeline(pipeline, nodeName, successors) {
+        const nodes = [];
+        pipeline.nodes.forEach((n) => {
+            if (n.nodeName === nodeName) {
+                n.cacheJobId = pipeline.rootJobId || pipeline.jobId; //eslint-disable-line
+            }
+            if (successors.includes(n.nodeName)) {
+                nodes.push(n);
+            }
+        });
+        return { ...pipeline, nodes };
+    }
+
+    _flatten(nodes, nodeName) {
+        const flatten = new Set();
+        flatten.add(nodeName);
+        if (nodes) {
+            nodes.forEach((n) => {
+                if (!flatten.has(n)) {
+                    flatten.add(n);
+                }
             });
         }
-        catch (e) {
-            error = e.response ? e.response.body.error : e.error;
+        return [...flatten];
+    }
+
+    async _getStoredExecution(jobId) {
+        const pipeline = await stateManager.getJobPipeline({ jobId });
+        if (!pipeline) {
+            throw new InvalidDataError(`unable to find pipeline ${jobId}`);
         }
-        return { error, pipeline };
+        return pipeline;
+    }
+
+    _findRelations(pipeline, nodeName) {
+        const graph = new DAG(pipeline, { validateNodesRelations: false });
+        const successorsMap = this._getSuccessors(graph, nodeName);
+        const successors = this._flatten(successorsMap, nodeName);
+        return { successors };
+    }
+
+    _getSuccessors(graph, nodeName, res = []) {
+        const successors = graph._childs(nodeName);
+        if (!successors) {
+            throw new InvalidDataError(`cant find relations for ${nodeName}`);
+        }
+        if (successors.length === 0) {
+            return null;
+        }
+        res.push(...successors);
+        successors.forEach(p => this._getSuccessors(graph, p, res));
+        return res;
     }
 }
