--- core/caching-service/.eslintignore
@@ -1,4 +1,0 @@
-coverage
-config
-tests
-bootstrap.js
