--- core/caching-service/app.js
@@ -1,6 +1,0 @@
-const bootstrap = require('./bootstrap');
-
-bootstrap.init().catch((error) => {
-    console.error(error);//eslint-disable-line
-    process.exit(1);
-});
