--- core/caching-service/bootstrap.js
@@ -1,55 +1,0 @@
-
-const Logger = require('@hkube/logger');
-const configIt = require('@hkube/config');
-const { main, logger } = configIt.load();
-const log = new Logger(main.serviceName, logger);
-const { componentName } = require('./lib/consts');
-const component = componentName.MAIN;
-
-const modules = [
-    './lib/api/rest',
-    './lib/runner'
-];
-
-class Bootstrap {
-    async init() {
-        try {
-            this._handleErrors();
-            log.info(`running application with env: ${configIt.env()}, version: ${main.version}, node: ${process.versions.node}`, { component });
-            await Promise.all(modules.map(m => require(m).init(main)));
-        }
-        catch (error) {
-            this._onInitFailed(error);
-        }
-    }
-
-    _onInitFailed(error) {
-        log.error(error.message, { component }, error);
-        log.error(error);
-        process.exit(1);
-    }
-
-    _handleErrors() {
-        process.on('exit', (code) => {
-            log.info(`exit code ${code}`, { component });
-        });
-        process.on('SIGINT', () => {
-            log.info('SIGINT', { component });
-            process.exit(0);
-        });
-        process.on('SIGTERM', () => {
-            log.info('SIGTERM', { component });
-            process.exit(0);
-        });
-        process.on('unhandledRejection', (error) => {
-            log.error(`unhandledRejection: ${error.message}`, { component }, error);
-            process.exit(1);
-        });
-        process.on('uncaughtException', (error) => {
-            log.error(`uncaughtException: ${error.message}`, { component }, error);
-            process.exit(1);
-        });
-    }
-}
-
-module.exports = new Bootstrap();
