--- core/caching-service/config/logger/config.development.js
@@ -1,9 +1,0 @@
-const config = {};
-config.transport = {
-    console: true,
-    fluentd: false,
-    logstash: false,
-    file: false
-};
-config.verbosityLevel = process.env.HKUBE_LOG_LEVEL || 2;
-module.exports = config;
