--- core/caching-service/config/logger/config.production.js
@@ -1,10 +1,0 @@
-const config = {};
-config.transport = {
-    console: false,
-    fluentd: true,
-    logstash: false,
-    file: false,
-    redis: true
-};
-config.verbosityLevel = process.env.HKUBE_LOG_LEVEL || 2;
-module.exports = config;
