--- core/caching-service/config/logger/config.test.js
@@ -1,9 +1,0 @@
-const config = {};
-config.transport = {
-    console: false,
-    fluentd: false,
-    logstash: false,
-    file: false
-};
-config.verbosityLevel = process.env.HKUBE_LOG_LEVEL || 2;
-module.exports = config;
