--- core/caching-service/config/main/config.production.js
@@ -1,2 +1,0 @@
-const config = {};
-module.exports = config;
