--- core/caching-service/config/main/config.test.js
@@ -1,3 +1,0 @@
-const config = {};
-config.algorithmType = process.env.ALGORITHM_TYPE || 'green-alg';
-module.exports = config;
