--- core/caching-service/lib/consts/component-name.js
@@ -1,5 +1,0 @@
-module.exports = {
-    API: 'API',
-    RUNNER: 'RUNNER',
-    DB: 'DB'
-};
