--- core/caching-service/lib/consts/index.js
@@ -1,3 +1,0 @@
-const componentName = require('./component-name');
-
-module.exports = { componentName };
