--- core/caching-service/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "caching-service",
-  "version": "2.1.2",
-  "description": "",
-  "main": "app.js",
-  "keywords": [],
-  "author": "Hkube",
-  "license": "MIT",
-  "dependencies": {
-    "@hkube/config": "^2.0.11",
-    "@hkube/dag": "^2.1.16",
-    "@hkube/db": "^1.0.32",
-    "@hkube/logger": "^1.0.27",
-    "@hkube/parsers": "^2.1.7",
-    "@hkube/rest-server": "^1.0.16",
-    "express": "^4.17.1"
-  },
-  "devDependencies": {
-    "body-parser": "^1.19.0",
-    "chai": "^4.2.0",
-    "clone": "^2.1.2",
-    "coveralls": "^3.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb-base": "^14.1.0",
-    "eslint-plugin-chai-friendly": "^0.5.0",
-    "eslint-plugin-import": "^2.20.2",
-    "mocha": "^8.0.1",
-    "nyc": "^15.1.0",
-    "sinon": "^9.0.2"
-  },
-  "scripts": {
-    "lint": "./node_modules/eslint/bin/eslint.js .",
-    "test": "NODE_ENV=test ./node_modules/mocha/bin/_mocha tests --colors --recursive --exit",
-    "test:cov": "NODE_ENV=test ./node_modules/nyc/bin/nyc.js --reporter=html --reporter=text --reporter=lcov mocha tests --colors --recursive --exit",
-    "coverage": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "start": "node app",
-    "build": "./dockerfile/build.sh caching-service"
-  }
-}
