--- core/caching-service/tests/test.js
@@ -1,19 +1,0 @@
-// const { graph } = require('../mocks/graph')
-// const NodesMap = require('../lib/create-graph')
-// const { expect } = require('chai');
-
-// let builtGraph = null;
-// xdescribe('root', () => {
-//     before(() => {
-
-//     });
-//     describe('test graph', () => {
-//         it('graph', () => {
-//             builtGraph = new NodesMap(graph.nodes);
-//             let pred = builtGraph.findDependentNodes(graph.nodes[1].nodeName);
-//             let pred2 = builtGraph.getDependentNodes("A");
-//             console.log(pred);
-//             expect(pred).to.not.equal(null);
-//         })
-//     })
-// });
